backend/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.7.0
azure-storage-blob==12.23.1
azure-ai-vision-imageanalysis==1.0.0
azure-ai-textanalytics==5.3.0
azure-identity==1.18.0
PyJWT==2.9.0
bcrypt==4.2.0
Pillow==11.0.0
python-multipart==0.0.18
